+++ spec/postgresql_adapter_spec.rb
@@ -25,4 +25,19 @@ describe ActiveRecord::ConnectionAdapters::PostgreSQLAdapter do
       expect(connection.executed_statements).to eq ['SET CONSTRAINTS "my_fk1", "my_fk2" DEFERRED']
     end
   end
+
+  describe "#defer_constraints" do
+    around do |example|
+      connection.dont_execute(&example)
+    end
+
+    it "defers and resets" do
+      block_called = false
+      connection.defer_constraints do
+        block_called = true
+      end
+      expect(block_called).to eq true
+      expect(connection.executed_statements).to eq ["SET CONSTRAINTS ALL DEFERRED", "SET CONSTRAINTS ALL IMMEDIATE"]
+    end
+  end
 end