+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add PostgreSQLAdapter#defer_constraints
+
 ## [0.1.0] - 2021-06-02
 
 - Initial release